Refactor connection status handling
Removed the `IsConnected` property from both `ConnectionService` and `IConnectionService`, which was previously used to represent the current connection status. The code that set this property to `true` upon establishing a connection has also been removed. Instead, an event `ConnectionRestored` is now invoked when a connection is established or restored, replacing the previous use of the `IsConnected` property.