Add GetSensorType() method to ISensorService interface
The ISensorService interface has been updated to include a new
method GetSensorType(). This method returns a string representing
the type of sensor, enhancing the interface by providing a way to
retrieve the sensor type. This can be useful for distinguishing
between different sensor implementations or for logging and
debugging purposes.