Improve thread stopping mechanism; remove unused handler
Enhanced the stopping mechanism of the sensor data thread by adding a check to log a message and initiate a forced stop if the thread does not stop within 5000 milliseconds. The thread is also set to `null` after attempting to stop it. Removed the `OnInternetRestored` event handler method, which previously restarted the service when the internet connection was restored.