Update library references to newer versions
Updated the HintPath for multiple references in the ESP32-NF-MQTT-DHT project to point to newer versions of the respective libraries. This includes upgrades for libraries such as `Iot.Device.Ahtxx`, `Iot.Device.Bmxx80`, `nanoFramework.Logging`, and others, reflecting a move from version 1.2.852 to 1.2.869 for several packages, along with similar incremental updates for others.